# interviewcalendar-model/src/main/java/pt/rt/interviewcalendar/model/dto/InterviewSlotDTO.java
package pt.rt.interviewcalendar.model.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

import java.util.Date;

@Builder
@Getter
@Setter
@ToString
@NoArgsConstructor
@AllArgsConstructor
@EqualsAndHashCode(onlyExplicitlyIncluded = true)
public class InterviewSlotDTO {

    @EqualsAndHashCode.Include
    private Date startDateTime;
    private Date endDateTime;
}